take out signin func in signin modal
