+++ src/components/Header/Icons/index.tsx
@@ -8,16 +8,12 @@ import { ReactComponent as User } from '@assets/svg/user.svg';
 import { Modal } from '@components/Modal';
 import { SingInModal } from '@components/SignInModal';
 import { paths } from '@constants/paths';
-import { useAppDispatch } from '@hooks/useAppDispatch';
 import { useAppSelector } from '@hooks/useAppSelector';
 import { useAuth } from '@hooks/useAuth';
-import { githubSignIn, googleSignIn } from '@store/reducers/authSlice';
-import { AuthProvider } from '@utils/providers';
 
 import './Icons.scss';
 
 export const Icons = memo(() => {
-  const dispatch = useAppDispatch();
   const { items } = useAppSelector((state) => state.cart);
   const { itemsInFavorite } = useAppSelector((state) => state.wishlist);
   const navigate = useNavigate();
@@ -39,25 +35,6 @@ export const Icons = memo(() => {
     } else {
       setIsModalActive(true);
     }
-  };
-
-  const handleSignIn = async (provider: AuthProvider) => {
-    switch (provider) {
-      case AuthProvider.GOOGLE:
-        await dispatch(googleSignIn());
-
-        break;
-      case AuthProvider.GITHUB:
-        await dispatch(githubSignIn());
-
-        break;
-
-      default:
-        break;
-    }
-
-    navigate(paths.wishlist);
-    setIsModalActive(false);
   };
 
   const quantityCart = items.length;
@@ -97,7 +74,10 @@ export const Icons = memo(() => {
           setIsModalActive={setIsModalActive}
           isModalActive={isModalActive}
         >
-          <SingInModal handleSignIn={handleSignIn} />
+          <SingInModal
+            setIsModalActive={setIsModalActive}
+            isWishlist
+          />
         </Modal>
       )}
     </div>
