+++ src/components/Modal/Modal.scss
@@ -13,7 +13,7 @@
   pointer-events: none;
   opacity: 0;
   transition: 0.3s linear;
-  z-index: 100;
+  z-index: 101;
   
   &__content {
     padding: 4px 10px;
